feat: Enable Swagger UI in Docker deployment environment
- Add SpringDoc configuration to application-docker.yml
- Configure Swagger UI to be accessible in production environment
- Set operations and tags sorting to alphabetical order
- Make Swagger settings configurable via environment variables
- Swagger will be available at http://[server]:8083/swagger-ui.html
