--- src/main/java/com/asyncsite/studyservice/domain/exception/StudyAlreadyApprovedException.java
@@ -1,11 +1,0 @@
-package com.asyncsite.studyservice.domain.exception;
-
-public class StudyAlreadyApprovedException extends RuntimeException {
-    public StudyAlreadyApprovedException(String message) {
-        super(message);
-    }
-    
-    public StudyAlreadyApprovedException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
