--- src/main/java/com/asyncsite/studyservice/domain/port/in/ProposeStudyUseCase.java
@@ -1,7 +1,0 @@
-package com.asyncsite.studyservice.domain.port.in;
-
-import com.asyncsite.studyservice.domain.model.Study;
-
-public interface ProposeStudyUseCase {
-    Study propose(String title, String description, String proposerId);
-}
