--- src/main/java/com/asyncsite/studyservice/domain/service/StudyNotFoundException.java
@@ -1,7 +1,0 @@
-package com.asyncsite.studyservice.domain.service;
-
-public class StudyNotFoundException extends RuntimeException {
-    public StudyNotFoundException(String message) {
-        super(message);
-    }
-}
