+++ src/main/java/com/asyncsite/studyservice/study/domain/exception/StudyAlreadyApprovedException.java
@@ -1,0 +1,11 @@
+package com.asyncsite.studyservice.study.domain.exception;
+
+public class StudyAlreadyApprovedException extends RuntimeException {
+    public StudyAlreadyApprovedException(String message) {
+        super(message);
+    }
+    
+    public StudyAlreadyApprovedException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
