+++ src/main/java/com/asyncsite/studyservice/study/domain/port/in/ManageStudyUseCase.java
@@ -1,0 +1,11 @@
+package com.asyncsite.studyservice.study.domain.port.in;
+
+import com.asyncsite.studyservice.study.domain.model.Study;
+
+import java.util.UUID;
+
+public interface ManageStudyUseCase {
+    Study approve(UUID studyId);
+    Study reject(UUID studyId);
+    Study terminate(UUID studyId);
+}
