+++ src/main/java/com/asyncsite/studyservice/study/domain/service/StudyNotFoundException.java
@@ -1,0 +1,7 @@
+package com.asyncsite.studyservice.study.domain.service;
+
+public class StudyNotFoundException extends RuntimeException {
+    public StudyNotFoundException(String message) {
+        super(message);
+    }
+}
