+++ src/main/resources/application-docker.yml
@@ -50,6 +50,18 @@ management:
     db:
       enabled: ${MANAGEMENT_HEALTH_DB_ENABLED:true}
 
+# SpringDoc/Swagger Configuration
+springdoc:
+  api-docs:
+    enabled: ${SPRINGDOC_API_DOCS_ENABLED:true}
+    path: ${SPRINGDOC_API_DOCS_PATH:/v3/api-docs}
+  swagger-ui:
+    enabled: ${SPRINGDOC_SWAGGER_UI_ENABLED:true}
+    path: ${SPRINGDOC_SWAGGER_UI_PATH:/swagger-ui.html}
+    operations-sorter: alpha
+    tags-sorter: alpha
+  show-actuator: ${SPRINGDOC_SHOW_ACTUATOR:false}
+
 # Logging Configuration
 logging:
   level:
